BottleShop/requirement.txt
asgiref==3.5.2
certifi==2022.6.15
charset-normalizer==2.1.1
Django==4.1
django-countries==7.3.2
django-jet==1.0.8
django-jet-4==0.0.1
django-paypal==2.0
idna==3.3
Pillow==9.2.0
pytz==2022.2.1
requests==2.28.1
six==1.16.0
sqlparse==0.4.2
stripe==4.1.0
typing-extensions==4.3.0
urllib3==1.26.12
